model/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
tensorflow==2.15.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
pillow==10.1.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
requests==2.31.0
pandas==2.1.3
python-multipart==0.0.6
pymongo==4.6.0
streamlit==1.28.2
python-dotenv==1.0.0
